LayoutTests/http/wpt/beacon/contentextensions/beacon-blocked-expected.txt

CONSOLE MESSAGE: line 9: Content blocker prevented frame displaying http://localhost:8800/CyberKit/beacon/contentextensions/beacon-blocked.html from loading a resource from http://localhost:8800/CyberKit/beacon/resources/beacon-preflight.py
CONSOLE MESSAGE: line 9: Beacon API cannot load http://localhost:8800/CyberKit/beacon/resources/beacon-preflight.py. Resource blocked by content blocker

PASS Content extensions should be able to block beacons 
